Fix AST headline nesting
Headlines are nested if their level is higher (more stars) than the current
headline. We're abusing the token.lvl field for this - as headlines can never
be indented we know the indentation must be 0 so we can cache the lvl (count of
stars) of the headline in that field.

This doesn't change anything right now so I'll postpone adding tests and stuff
until there are actual use cases for the AST and stuff.