Emit some-descriptive-id instead of headline-1.
I want to my hugo site with org-mode files to support links such as

  /page#meaningful-anchor

This commit is pretty dumb because it will transform

  ** [[https://github.com/plujon/go-org][go-org]]

to

  <h2 id="https-github-com-plujon-go-org-go-org">

Other projects I know of that provide similar transformations are
gollum-lib and AnchorJS.  The former uses nokogiri (an html parser) to
strip tags before applying the transformation, and the latter uses the
browser's el.textContent().

Unfortunately, I'm not a go-programmer and I don't know how to iterate
over the parsed nodes to do this in go-org in a reasonable amount of
time.

Possibly useful referneces:

https://github.com/gollum/gollum-lib/blob/f32a59d6e7c489e0f7a735987272fc23211e8092/lib/gollum-lib/filter/toc.rb#L109

https://github.com/bryanbraun/anchorjs/blob/e0f62ebb48c386c8aa3cf7a47cced5ef17f60e3c/anchor.js#L239